data_flow: Adds to CHRE build
Adds data flows to the CHRE build. This requires adding a copy of the
AIDL-generated SharedDataRegion.h header containing the ABI for data
flows to be used where AIDL code-generation is not supported.

The new dependencies are only enabled if CHRE_DATA_FLOW_SUPPORT_ENABLED
is true.

Bug: 475656750
Test: atest --host contexthub_data_flow_unit_tests. CHRE builds for
enabled platforms.
BYPASS_LARGE_CHANGE_WARNING: Adding generated header.

+++ data_flow/queue.cc
@@ -1458,7 +1458,8 @@ pw::Status VariableDataProducer::push(pw::ConstByteSpan element) {
 void VariableDataProducer::updateFirstElementIndex() {
   auto *tailBlock = internal::fromOffset<internal::VariableDataBlock>(
       mRegion, mDesc->tailBlockOffsetBytes, kBlockLayout);
-  if (tailBlock->header.firstElementIndex == kBlockCapacity) {
+  if (static_cast<uint32_t>(tailBlock->header.firstElementIndex) ==
+      kBlockCapacity) {
     // Only set the first element index if this is the first variable size
     // element to be written into this block (on this pass through the block).
     tailBlock->header.firstElementIndex =
